+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to Cordyceps will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
-## [Unreleased]
+## [1.4.12] - 2026-06-24
 
 ### Fixed